SKALED-1583 Remove some patch files
